add module config
- add in config MetaConfig class
- use fabric pattern to instance MetaConfig object
- add time check of running script
