+++ __init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+"""
+Created on 19.08.2017
+
+@author: NuclearRazor
+"""
+
